refactor: 레포지토리 추상화 및 패키지 변경
