+++ src/main/java/roomescape/core/domain/ReservationTime.java
@@ -1,0 +1,47 @@
+package roomescape.core.domain;
+
+import java.time.LocalTime;
+import java.util.Objects;
+
+public class ReservationTime {
+
+    private Long id;
+    private LocalTime startAt;
+
+    protected ReservationTime() {
+    }
+
+    public ReservationTime(LocalTime startAt) {
+        this(null, startAt);
+    }
+
+    public ReservationTime(Long id, LocalTime startAt) {
+        this.id = id;
+        this.startAt = startAt;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public LocalTime getStartAt() {
+        return startAt;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        ReservationTime that = (ReservationTime) o;
+        return Objects.equals(id, that.id);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hashCode(id);
+    }
+}
