+++ src/main/java/roomescape/core/service/dto/ReservationTimeServiceResponse.java
@@ -1,0 +1,14 @@
+package roomescape.core.service.dto;
+
+import java.time.LocalTime;
+import roomescape.core.domain.ReservationTime;
+
+public record ReservationTimeServiceResponse(
+        Long id,
+        LocalTime startAt
+) {
+
+        public static ReservationTimeServiceResponse from(ReservationTime reservationTime) {
+                return new ReservationTimeServiceResponse(reservationTime.getId(), reservationTime.getStartAt());
+        }
+}
