--- src/main/java/roomescape/domain/Reservation.java
@@ -1,66 +1,0 @@
-package roomescape.domain;
-
-import java.time.LocalDate;
-import java.util.Objects;
-
-public class Reservation {
-
-    private Long id;
-    private String name;
-    private LocalDate date;
-    private ReservationTime time;
-
-    protected Reservation() {
-    }
-
-    public Reservation(String name, LocalDate date, ReservationTime time) {
-        this.name = name;
-        this.date = date;
-        this.time = time;
-    }
-
-    public Reservation(Long id, String name, LocalDate date, ReservationTime time) {
-        this.id = id;
-        this.name = name;
-        this.date = date;
-        this.time = time;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public LocalDate getDate() {
-        return date;
-    }
-
-    public Long getTimeId() {
-        return time.getId();
-    }
-
-    public ReservationTime getTime() {
-        return time;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        Reservation that = (Reservation) o;
-        return Objects.equals(id, that.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hashCode(id);
-    }
-}
-
