+++ src/main/java/roomescape/web/controller/AdminController.java
@@ -1,0 +1,25 @@
+package roomescape.web.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/admin")
+public class AdminController {
+
+    @GetMapping
+    public String admin() {
+        return "admin/index";
+    }
+
+    @GetMapping("/reservation")
+    public String reservation() {
+        return "admin/reservation";
+    }
+
+    @GetMapping("/time")
+    public String time() {
+        return "admin/time";
+    }
+}
