+++ src/main/java/roomescape/web/controller/HomeController.java
@@ -1,0 +1,13 @@
+package roomescape.web.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class HomeController {
+
+    @GetMapping
+    public String home() {
+        return "redirect:/admin";
+    }
+}
