+++ src/main/java/roomescape/web/controller/dto/ReservationTimeResponse.java
@@ -1,0 +1,16 @@
+package roomescape.web.controller.dto;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import java.time.LocalTime;
+import roomescape.core.service.dto.ReservationTimeServiceResponse;
+
+public record ReservationTimeResponse(
+        Long id,
+        @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "HH:mm", timezone = "Asia/Seoul")
+        LocalTime startAt
+) {
+
+        public static ReservationTimeResponse from(ReservationTimeServiceResponse reservationTimeServiceResponse) {
+                return new ReservationTimeResponse(reservationTimeServiceResponse.id(), reservationTimeServiceResponse.startAt());
+        }
+}
